app/db_conn.py
from dotenv import load_dotenv
import os

load_dotenv()


import psycopg2

def get_connection():
    try:
        return psycopg2.connect(
            host=os.getenv("DB_HOST"),
            database=os.getenv("DB_NAME"),
            user=os.getenv("DB_USER"),
            password=os.getenv("DB_PASSWORD"),
            port=os.getenv("DB_PORT")
        )
    except Exception as err:
        print("DB connection error:", err)
        return None