app/requirements.txt
requests 
pandas 
openpyxl
psycopg2-binary
streamlit
plotly
matplotlib 
seaborn
python-dotenv
streamlit
